requirements-gcp.txt
# Core ML and Data Science
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# MLflow and Experiment Tracking
mlflow==2.9.2

# GCP SDK (already installed, but ensure versions)
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.1

# DVC with GCS support
dvc==3.30.1
dvc-gs==2.22.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
